Move ActivitySearch
