+++ ui/src/components/activity/ActivitySearch/connect.js
